refactor: extract subagent schemas and notification helpers
